requirements.txt
requests
Pillow
opencv-python
numpy
Flask